.history/requirements_20240810144742.txt
pickle5
streamlita